requirements.txt
PyDrive==1.3.1
pandas==1.0.5
python_docx==0.8.10
google_auth_oauthlib==0.4.1
pygdrive3==0.7.0
airtable_python_wrapper==0.13.0
airtable==0.3.1
docx==0.2.4
docx2pdf==0.1.7
google_api_python_client==1.10.0
protobuf==3.12.2
xlrd==1.2.0